requirements.txt
flask
yt-dlp
git+https://github.com/openai/whisper.git
more-itertools
numba
numpy
tiktoken
tqdm
openai
python-dotenv